requirements.txt
pandas
python-dotenv
langchain-google-genai
langchain
tqdm
scikit-learn